Refactor: Moved Constants and renamed launchSdk
- Moved Constants object to domain.util package.
- Renamed launchSdk to initiate in EudiRQESUi.
- Updated SuccessScreen to not show any navigation action in the topbar.
