+++ rqes-ui-sdk/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?><!--
+<?xml version="1.0" encoding="utf-8"?>
+<!--
   ~ Copyright (c) 2023 European Commission
   ~
   ~ Licensed under the EUPL, Version 1.2 or - as soon they will be approved by the European
@@ -15,36 +16,20 @@
   -->
 
 <resources>
-
-    <!-- Sign Document Errors -->
-    <string name="sign_document_generic_error">Something went wrong while signing your document(s)</string>
-
-
-    <string name="sign_document_title">Sign a document</string>
-    <string name="sign_document_subtitle">Please confirm signing of the following</string>
-    <string name="generic_error_description">If the issue persists, please contact customer support</string>
-
-    <string name="sign_document_bottom_sheet_cancel_confirmation_title">Cancel signing process?</string>
-    <string name="sign_document_bottom_sheet_cancel_confirmation_subtitle">Cancel wil redirect you back to the documents list without signing your document</string>
-    <string name="sign_document_bottom_sheet_continue_button_text">Continue signing</string>
-    <string name="sign_document_bottom_sheet_cancel_button_text">Cancel signing</string>
-    <string name="sign_document_pdf_selection_item_label">VIEW</string>
-
-    <string name="sign_document_qtsp_title">Select remote signing service</string>
-    <string name="sign_document_qtsp_subtitle">Remote signing enables you to digitally sign documents
-        without the need for locally installed digital identities. Cloud-hosted signing service makes
-        remote signing possible.</string>
-
-    <!-- Generic -->
-    <string name="generic_error_button_retry">TRY AGAIN</string>
-    <string name="generic_error_message">Oups! Something went wrong</string>
-
-    <string name="generic_sign_button_text">Sign</string>
-
     <!-- Content Description -->
     <string name="content_description_arrow_back_icon">Go Back</string>
     <string name="content_description_close_icon">Close</string>
     <string name="content_description_handle_bar">Handle bar</string>
     <string name="content_description_verified_icon">Verified</string>
     <string name="content_description_more_vert_icon">More Items</string>
+
+    <!-- Generic -->
+    <string name="generic_error_button_retry">TRY AGAIN</string>
+    <string name="generic_error_message">Oups! Something went wrong</string>
+    <string name="generic_error_description">If the issue persists, please contact customer support</string>
+    <string name="generic_error_document_not_found">No Document data found</string>
+    <string name="generic_error_qtsp_not_found">No selected QTSP found</string>
+
+    <!-- Sign Document Errors -->
+    <string name="sign_document_generic_error">Something went wrong while signing your document(s)</string>
 </resources>
